+++ includes/ObserverPattern.h
@@ -8,14 +8,14 @@ namespace ObserverPattern{
 	// to use this class for observer pattern, you must define a new class 
 	// to extend this class, and then implement the abstract member
 	// method `updateData()`.
-	class myAbstractObserver{
+	class myObserverInterface{
 		// public methods
 		public:
 			// Description: Update info
 			// @param sKey: The keyword of the data
 			// @param pData: The data is sent to observer
 			virtual void updateData(const string sKey, void *pData) = 0;
-	};	// End of class myAbstractObserver
+	};	// End of class myObserverInterface
 
 	// Description:
 	// This abstract class is to send data to all observers. If you want
@@ -25,19 +25,19 @@ namespace ObserverPattern{
 		// protected fields
 		protected:
 			// The list is to store all observer.
-			vector<myAbstractObserver*> m_poObserver;
+			vector<myObserverInterface*> m_poObserver;
 
 		public:
 			mySubjectInterface();
 			~mySubjectInterface();
 
 			// Description: Register observer into the list
-			// @param poObserver: The object pointer of the class myAbstractObserver
-			void registerObserver(myAbstractObserver *poObserver);
+			// @param poObserver: The object pointer of the class myObserverInterface
+			void registerObserver(myObserverInterface *poObserver);
 
 			// Description: Remove the observer from the list
-			// @param poObserver: The object pointer of the class myAbstractObserver
-			void removeObserver(myAbstractObserver *poObserver);
+			// @param poObserver: The object pointer of the class myObserverInterface
+			void removeObserver(myObserverInterface *poObserver);
 
 			// Description: Notify the data to observers
 			// @param sKey: The keyword of the data